TRIX-434 renaming templates_dir into templates_directory and templatelist into templates_list

+++ daemon/common/constant.py
@@ -216,7 +216,7 @@ CONSTANT = {
     'SERVICES': {'DHCP': None, 'DNS': None, 'CONTROL': None, 'COOLDOWN': None, 'COMMAND': None},
     'DHCP': {'OMAPIKEY': None},
     'BMCCONTROL': {'BMC_BATCH_SIZE': None, 'BMC_BATCH_DELAY': None},
-    'TEMPLATES': {'TEMPLATES_DIR': None, 'TEMPLATELIST': None,  'TMP_DIRECTORY': None}
+    'TEMPLATES': {'TEMPLATES_DIRECTORY': None, 'TEMPLATES_LIST': None,  'TMP_DIRECTORY': None}
 }
 
 if check_path_state(CONFIGFILE):
@@ -226,7 +226,7 @@ else:
     LOGGER.error(f'Unable to get configurations from {CONFIGFILE} file')
 
 
-## Sanity Checks On LOGFILE, IMAGE_FILES, TEMPLATES_DIR, TEMPLATELIST, KEYFILE
+## Sanity Checks On LOGFILE, IMAGE_FILES, TEMPLATES_DIRECTORY, TEMPLATES_LIST, KEYFILE
 
 # the log path, not the file. we can create the file very well by ourselves... - Antoine aug 11 2023
 logpath = os.path.dirname(CONSTANT['LOGGER']['LOGFILE'])
@@ -235,8 +235,8 @@ sanitize = [
 #                CONSTANT['LOGGER']['LOGFILE'],
                 logpath,
                 CONSTANT['FILES']['IMAGE_FILES'],
-                CONSTANT['TEMPLATES']['TEMPLATES_DIR'],
-                CONSTANT['TEMPLATES']['TEMPLATELIST'],
+                CONSTANT['TEMPLATES']['TEMPLATES_DIRECTORY'],
+                CONSTANT['TEMPLATES']['TEMPLATES_LIST'],
                 CONSTANT['PLUGINS']['PLUGINS_DIR'],
                 CONSTANT['FILES']['KEYFILE']
             ]
@@ -254,25 +254,26 @@ with open(CONSTANT['FILES']['KEYFILE'], 'r', encoding='utf-8') as key_file:
     LUNAKEY = key_file.read()
     LUNAKEY = LUNAKEY.replace('\n', '')
 
-with open(CONSTANT['TEMPLATES']['TEMPLATELIST'], 'r', encoding='utf-8') as template_json:
+with open(CONSTANT['TEMPLATES']['TEMPLATES_LIST'], 'r', encoding='utf-8') as template_json:
     data = json.load(template_json)
 if 'files' in data.keys():
-    if len(CONSTANT["TEMPLATES"]["TMP_DIRECTORY"]) > 1:
+    if len(CONSTANT["TEMPLATES"]["TMP_DIRECTORY"]) > 1 and os.path.exists(CONSTANT["TEMPLATES"]["TMP_DIRECTORY"]):
         shutil.rmtree(CONSTANT["TEMPLATES"]["TMP_DIRECTORY"])
     if not os.path.exists(CONSTANT["TEMPLATES"]["TMP_DIRECTORY"]):
         os.makedirs(CONSTANT["TEMPLATES"]["TMP_DIRECTORY"])
     if not os.path.exists(CONSTANT["TEMPLATES"]["TMP_DIRECTORY"]):
-        LOGGER.error(f"Cannot create directory {CONSTANT["TEMPLATES"]["TMP_DIRECTORY"]}")
+        LOGGER.error(f"Cannot create directory {CONSTANT['TEMPLATES']['TMP_DIRECTORY']}")
     else:
-        if check_path_state(f'{CONSTANT["TEMPLATES"]["TEMPLATES_DIR"]}/{templatefiles}'):
-            for templatefiles in data['files']:
-                copy_source = f'{CONSTANT["TEMPLATES"]["TEMPLATES_DIR"]}/{templatefiles}'
-                shutil.copyfile(copy_source, CONSTANT["TEMPLATES"]["TMP_DIRECTORY"])
-        else:
-            error_msg = f'{CONSTANT["TEMPLATES"]["TEMPLATES_DIR"]}/{templatefiles} is not writable'
-            LOGGER.error(error_msg)
+        for templatefiles in data['files']:
+            if check_path_state(f'{CONSTANT["TEMPLATES"]["TEMPLATES_DIRECTORY"]}/{templatefiles}'):
+                copy_source = f'{CONSTANT["TEMPLATES"]["TEMPLATES_DIRECTORY"]}/{templatefiles}'
+                copy_destination = f'{CONSTANT["TEMPLATES"]["TMP_DIRECTORY"]}/{templatefiles}'
+                shutil.copyfile(copy_source, copy_destination)
+            else:
+                error_msg = f'{CONSTANT["TEMPLATES"]["TEMPLATES_DIRECTORY"]}/{templatefiles} is not writable'
+                LOGGER.error(error_msg)
 else:
-    LOGGER.error(f'{CONSTANT["TEMPLATES"]["TEMPLATELIST"]} have no files')
+    LOGGER.error(f'{CONSTANT["TEMPLATES"]["TEMPLATES_LIST"]} have no files')
